feat(docked-window): improved window management
Separated os.ui.window utilities from os.ui.WindowUI. Updated CSS and styles for docked windows.
