+++ views/window/window.html
@@ -1,4 +1,4 @@
-<div class="js-window modal d-flex c-window o-window" ng-class="active && 'active shadow' || 'shadow-sm'">
+<div class="js-window modal d-flex c-window o-window" ng-class="{'active shadow': active && !dock, 'shadow-sm': !active && !dock}">
   <div class="modal-content d-flex flex-column flex-fill" ng-class="{'bg-transparent': overlay, 'border-0': !border}">
     <div class="js-window__header modal-header d-flex justify-content-between flex-shrink-0"
         ng-class="headerClass"
